Fix: fixing and re-organizing gguf tests
