+++ tests/brevitas_examples/gguf/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2026, Advanced Micro Devices, Inc. All rights reserved.
+# SPDX-License-Identifier: BSD-3-Clause
